requirements.txt
Flask==3.0.2
pg8000==1.30.5
python-dotenv==1.0.1
gunicorn==21.2.0
pywebpush==1.14.0
Pillow==10.2.0
pypdf==4.0.1

